Create main.c
I am new to coding and have a limited knowledge of the subject. I have a question for someone more experienced than me. In this lab we are supposed to have the the factors for the number be right biased, only 3 factors per line and all commas must align. It works up to the point where i need the alignment to be proper. (I will upload a picture of what the output should look like)

Also feel free to help my improve my code in other ways, if you see something say it! Its the only way to learn.